Move code from lib.rs to unit.rs
